refactor(hook): 카드목록 페이지에서 data fetch 로직을 커스텀 훅으로 분리
